Try to merge sync and reactive tests
